refactor: split gui and pipeline
