--- dexi_gui.py
@@ -1,6 +1,0 @@
-"""Entry point for the Dexi LineArt GUI."""
-
-from src.dexi_gui import main
-
-if __name__ == "__main__":
-    main()
